+++ docs/html/search/variables_1.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['savecurrentposition_0',['saveCurrentPosition',['../do__text__text_8php.html#a7be69b326841ebf3576e44dc906046a3',1,'do_text_text.php']]]
-];