+++ edit_tword.php
@@ -163,8 +163,8 @@ else {  // if (! isset($_REQUEST['op']))
     $scrdir = getScriptDirectionTag($lang);
 
     ?>
-<script type="text/javascript" src="js/unloadformcheck.js" charset="utf-8"></script>
 <script type="text/javascript">
+    $(document).ready(ask_before_exiting);
     $(window).on('beforeunload',function() {
         setTimeout(function() {window.parent.frames['ru'].location.href = 'empty.html';}, 0);
     });