+++ inc/ajax_update_media_select.php
@@ -19,7 +19,8 @@ require_once __DIR__ . '/session_utility.php';
 /**
  * Change the current working directory and find media path
  */
-function do_ajax_update_media_select() {
+function do_ajax_update_media_select()
+{
     chdir('..');
     return selectmediapath('TxAudioURI');
 }