+++ settings.php
@@ -196,7 +196,9 @@ if (isset($_REQUEST['op'])) {
 echo error_message_with_hide($message, 1);
 
 ?>
-<script type="text/javascript" src="js/unloadformcheck.js" charset="utf-8"></script>
+<script type="text/javascript" charset="utf-8">
+    $(document).ready(ask_before_exiting);
+</script>
 <form class="validate" action="<?php echo $_SERVER['PHP_SELF']; ?>" method="post">
 <table class="tab3" cellspacing="0" cellpadding="5">
 <!-- ******************************************************* -->